+++ functions.py
@@ -6,6 +6,11 @@ def minus(a, b):
     """Subtract b from a."""
     return a - b
 
+def exp(a, b):
+    """Exponential a**b."""
+    return a ** b
+
 def divide(a, b):
     """Divide b from a"""
     return a / b
+