+++ CHANGELOG.md
@@ -16,4 +16,7 @@
 
 ### Added
 
+
+- Added `exp(a, b)` function in `functions.py` to perform exponential.
+=======
 - Added `divide(a, b)` function in `functions.py` to perform division.